Begin refactoring for hibernate & guice
