+++ build.gradle
@@ -36,7 +36,7 @@ repositories {
 
 dependencies {
     shadow 'org.spongepowered:spongeapi:7.0.0-SNAPSHOT'
-    shadow 'com.github.Atherys-Horizons:AtherysCore:master-SNAPSHOT'
+    shadow 'com.github.Atherys-Horizons:AtherysCore:1.13.2'
     compile 'com.github.SimonFlash.TeslaPowered:TeslaLibs:v1.1.5'
 }
