+++ src/main/java/com/atherys/roleplay/AtherysRoleplayConfig.java
@@ -1,6 +1,5 @@
 package com.atherys.roleplay;
 
-import com.atherys.core.database.mongo.MongoDatabaseConfig;
 import com.atherys.core.utils.PluginConfig;
 import com.atherys.roleplay.cards.Nation;
 import ninja.leaping.configurate.objectmapping.Setting;
@@ -20,9 +19,6 @@ public class AtherysRoleplayConfig extends PluginConfig {
 
     @Setting("dialog_msg_delay")
     public long DIALOG_MESSAGE_DELAY = 2;
-
-    @Setting("database")
-    public MongoDatabaseConfig DATABASE = new MongoDatabaseConfig();
 
     @Setting("maximum_age")
     public int MAXIMUM_AGE = 150;
