+++ src/main/java/com/atherys/roleplay/command/card/CreateCardCommand.java
@@ -1,8 +1,7 @@
-package com.atherys.roleplay.commands.card;
+package com.atherys.roleplay.command.card;
 
 import com.atherys.core.command.annotation.Aliases;
 import com.atherys.core.command.annotation.Description;
-import com.atherys.roleplay.CardManager;
 import org.spongepowered.api.command.CommandException;
 import org.spongepowered.api.command.CommandResult;
 import org.spongepowered.api.command.CommandSource;
