+++ src/main/java/com/atherys/roleplay/command/card/set/SetCardDescriptionCommand.java
@@ -1,9 +1,8 @@
-package com.atherys.roleplay.commands.card.set;
+package com.atherys.roleplay.command.card.set;
 
 import com.atherys.core.command.ParameterizedCommand;
 import com.atherys.core.command.annotation.Aliases;
 import com.atherys.core.command.annotation.Description;
-import com.atherys.roleplay.AtherysRoleplay;
 import com.atherys.roleplay.CardManager;
 import com.atherys.roleplay.RoleplayMsg;
 import org.spongepowered.api.command.CommandException;
@@ -18,32 +17,28 @@ import org.spongepowered.api.text.Text;
 import javax.annotation.Nonnull;
 import java.util.Optional;
 
-@Aliases("name")
-@Description("Sets the name your character.")
-public class SetCardNameCommand implements ParameterizedCommand {
+@Aliases("description")
+@Description("Sets the description of your character. Subsequent uses will add to your description.")
+public class SetCardDescriptionCommand implements ParameterizedCommand {
 
     @Nonnull
     @Override
     public CommandResult execute(@Nonnull CommandSource src, @Nonnull CommandContext args) throws CommandException {
         if(!(src instanceof Player)) return CommandResult.empty();
-
         Player player = (Player) src;
-        Optional<String> name = args.getOne("name");
-        name.ifPresent(n -> {
-            if(n.length() > AtherysRoleplay.getConfig().MAXIMUM_NAME_LENGTH) {
-                RoleplayMsg.error(player, "Your name must be under " + AtherysRoleplay.getConfig().MAXIMUM_NAME_LENGTH);
-            } else {
-                CardManager.getInstance().getCard(player).setName(n);
-                RoleplayMsg.info(player, "Character name set.");
-            }
+        Optional<String> description = args.getOne("description");
+        description.ifPresent(desc -> {
+            RoleplayMsg.info(player, "Character description updated.");
+            CardManager.getInstance().getCard(player).addDescription(" " + desc);
         });
+
         return CommandResult.success();
     }
 
     @Override
     public CommandElement[] getArguments() {
         return new CommandElement[]{
-                GenericArguments.remainingJoinedStrings(Text.of("name"))
+                GenericArguments.remainingJoinedStrings(Text.of("description"))
         };
     }
 }
