+++ src/main/java/com/atherys/roleplay/command/card/set/SetCardNameCommand.java
@@ -1,12 +1,10 @@
-package com.atherys.roleplay.commands.card.set;
+package com.atherys.roleplay.command.card.set;
 
 import com.atherys.core.command.ParameterizedCommand;
 import com.atherys.core.command.annotation.Aliases;
 import com.atherys.core.command.annotation.Description;
 import com.atherys.roleplay.AtherysRoleplay;
-import com.atherys.roleplay.CardManager;
 import com.atherys.roleplay.RoleplayMsg;
-import com.atherys.roleplay.cards.Nation;
 import org.spongepowered.api.command.CommandException;
 import org.spongepowered.api.command.CommandResult;
 import org.spongepowered.api.command.CommandSource;
@@ -17,12 +15,11 @@ import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.text.Text;
 
 import javax.annotation.Nonnull;
-import java.util.HashMap;
 import java.util.Optional;
 
-@Aliases("nation")
-@Description("Sets the nation of your character.")
-public class SetCardNationCommand implements ParameterizedCommand {
+@Aliases("name")
+@Description("Sets the name your character.")
+public class SetCardNameCommand implements ParameterizedCommand {
 
     @Nonnull
     @Override
@@ -30,28 +27,22 @@ public class SetCardNationCommand implements ParameterizedCommand {
         if(!(src instanceof Player)) return CommandResult.empty();
 
         Player player = (Player) src;
-        Optional<Nation> name = args.getOne("nation");
+        Optional<String> name = args.getOne("name");
         name.ifPresent(n -> {
-            CardManager.getInstance().getCard(player).setNationality(n.getName());
-            RoleplayMsg.nationMessage(player, n);
+            if(n.length() > AtherysRoleplay.getConfig().MAXIMUM_NAME_LENGTH) {
+                RoleplayMsg.error(player, "Your name must be under " + AtherysRoleplay.getConfig().MAXIMUM_NAME_LENGTH);
+            } else {
+                CardManager.getInstance().getCard(player).setName(n);
+                RoleplayMsg.info(player, "Character name set.");
+            }
         });
         return CommandResult.success();
     }
 
-    private static HashMap<String, Nation> getChoices(){
-        HashMap<String, Nation> choices = new HashMap<>();
-        AtherysRoleplay.getConfig().NATIONS.forEach(nation -> choices.put(nation.getName(), nation));
-        return choices;
-    }
-
-    private HashMap<String, Nation> choices = getChoices();
-
     @Override
     public CommandElement[] getArguments() {
         return new CommandElement[]{
-                GenericArguments.choices(Text.of("nation"), choices)
+                GenericArguments.remainingJoinedStrings(Text.of("name"))
         };
     }
 }
-
-
