+++ src/main/java/com/atherys/roleplay/command/chat/ShoutChatCommand.java
@@ -1,0 +1,4 @@
+package com.atherys.roleplay.command.chat;
+
+public class ShoutChatCommand {
+}
