--- src/main/java/com/atherys/roleplay/commands/chat/ShoutChatCommand.java
@@ -1,4 +1,0 @@
-package com.atherys.roleplay.commands.chat;
-
-public class ShoutChatCommand {
-}
