--- src/main/java/com/atherys/roleplay/commands/chat/TalkChatCommand.java
@@ -1,4 +1,0 @@
-package com.atherys.roleplay.commands.chat;
-
-public class TalkChatCommand {
-}
