--- src/main/java/com/atherys/roleplay/commands/chat/WhisperChatCommand.java
@@ -1,4 +1,0 @@
-package com.atherys.roleplay.commands.chat;
-
-public class WhisperChatCommand {
-}
