+++ src/main/java/com/atherys/roleplay/service/CardService.java
@@ -1,0 +1,8 @@
+package com.atherys.roleplay.service;
+
+import com.google.inject.Singleton;
+
+@Singleton
+public class CardService {
+
+}
